Move tests to the separate files
