+++ tests/routers/api/test_packets.py
@@ -1,0 +1,34 @@
+from fastapi import status
+from fastapi.testclient import TestClient
+
+from ph.main import app
+from ph.models.creator_packets import CreatorPacketsRequest, CreatorPacketsResponse
+from ph.models.decoded_hex import DecodedHexResponse
+
+client = TestClient(app)
+
+
+class TestAPIPackets:
+    @staticmethod
+    def test_post_api_packets__success():
+        response = client.post(
+            "/api/packets",
+            json=CreatorPacketsRequest(packets=["Ether"]).dict(),
+        )
+        assert response.status_code == status.HTTP_201_CREATED
+        json_response = CreatorPacketsResponse(**response.json())
+        assert len(json_response.packets) == 1
+        assert json_response.packets[0]["Ethernet"]
+        assert len(json_response.packets[0]["Ethernet"]) == 3
+
+    @staticmethod
+    def test_post_api_packets__negative__packet_not_exists_in_scapy():
+        response = client.post(
+            "/api/packets",
+            json=CreatorPacketsRequest(packets=["NonExistingPacket"]).dict(),
+        )
+        assert response.status_code == status.HTTP_500_INTERNAL_SERVER_ERROR
+        assert (
+            response.json()["detail"]["error"]
+            == "Layer is not supported 'NonExistingPacket'"
+        )
